requirements.txt
Flask==2.1.2
python-webpack-boilerplate==1.0.0
boto3==1.22.13
python-dotenv==0.20.0
gunicorn==20.1.0